pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-accessibility_auditor"
version = "0.1.0"
description = "Accessibility Auditor environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "playwright>=1.40.0",
    "axe-playwright-python>=0.1.7",
    "fastapi>=0.100.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m accessibility_auditor.server.app
server = "accessibility_auditor.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["accessibility_auditor", "accessibility_auditor.server"]
package-dir = { "accessibility_auditor" = ".", "accessibility_auditor.server" = "server" }